# api-schema/api.yaml
openapi: 3.0.0
servers:
  - url: 'http://petstore.swagger.io/v2'
info:
  description: >-
    This APIs for iwf SDKs to operate workflows
  version: 1.0.0
  title: Workflow APIs
  license:
    name: Apache-2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
paths:
...
  /api/v1/workflow/show:
    post:
      summary: "show details for a workflow execution"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowShowRequest'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowShowResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
components:
  schemas:
    KeyValue:
      type: object
      properties:
        key:
          type: string
        value:
          $ref: '#/components/schemas/EncodedObject'
    EncodedObject:
      type: object
      properties:
        encoding:
          type: string
        data:
          type: string
...
    SearchAttributeValueType: # see https://docs.temporal.io/concepts/what-is-a-search-attribute#types
      type: string
      enum:
        - KEYWORD
        - TEXT
        - DATETIME
        - INT
        - DOUBLE
        - BOOL
        - KEYWORD_ARRAY # NOTE: for Cadence/Temporal keyword array is registered in the same type as KEYWORD
    SearchAttribute:
      type: object
      properties:
        key:
          type: string
        stringValue: # It's used for any value than can be stored as string: keyword, text, and DATETIME
          type: string
        integerValue: # It's used for INT type
          type: integer
          format: int64
        doubleValue: # For DOUBLE type
          type: number
          format: double
        boolValue: # for BOOL type
          type: boolean
        stringArrayValue: # for any type than can be stored as string array: KEYWORD, TEXT, and DATETIME
          type: array
          items:
            type: string
        valueType:
          $ref: '#/components/schemas/SearchAttributeValueType'
    RetryPolicy:
      type: object
      properties:
        initialIntervalSeconds: # default to 1s
          type: integer
        backoffCoefficient: # default to 2
          type: number
          format: float
        maximumIntervalSeconds: # default to 100x of initial interval
          type: integer
        maximumAttempts: # default to 0, means unlimited
          type: integer
        maximumAttemptsDurationSeconds: # the end to end duration of all attempts
          type: integer
...
    ExecuteApiFailurePolicy:
      type: string
      enum:
        - FAIL_WORKFLOW_ON_EXECUTE_API_FAILURE
        - PROCEED_TO_CONFIGURED_STATE
    StartApiFailurePolicy:
      type: string
      enum:
        - FAIL_WORKFLOW_ON_START_API_FAILURE
        - PROCEED_TO_DECIDE_ON_START_API_FAILURE # next states derived from decide() API
    WaitUntilApiFailurePolicy:
      type: string
      enum:
        - FAIL_WORKFLOW_ON_FAILURE
        - PROCEED_ON_FAILURE # next states derived from execute() API
    WorkflowStateOptions:
      type: object
      properties:
        searchAttributesLoadingPolicy:
          $ref: '#/components/schemas/PersistenceLoadingPolicy'
        waitUntilApiSearchAttributesLoadingPolicy:
          $ref: '#/components/schemas/PersistenceLoadingPolicy'
        executeApiSearchAttributesLoadingPolicy:
          $ref: '#/components/schemas/PersistenceLoadingPolicy'
        dataObjectsLoadingPolicy:
          $ref: '#/components/schemas/PersistenceLoadingPolicy'
        startApiTimeoutSeconds: # NOTE: this is the timeout for the single attempt of Cadence/Temporal activity(start to close timeout)
          type: integer
        decideApiTimeoutSeconds: # NOTE: this is the timeout for the single attempt of Cadence/Temporal activity(start to close timeout)
          type: integer
        startApiRetryPolicy:
          $ref: '#/components/schemas/RetryPolicy'
        decideApiRetryPolicy:
          $ref: '#/components/schemas/RetryPolicy'
        startApiFailurePolicy:
          $ref: '#/components/schemas/StartApiFailurePolicy'
        executeApiFailurePolicy:
          $ref: '#/components/schemas/ExecuteApiFailurePolicy'
        executeApiFailureProceedStateId:
          type: string
        executeApiFailureProceedStateOptions:
          $ref: '#/components/schemas/WorkflowStateOptions'
        skipStartApi:
          # all below are for compatibility
          type: boolean
        waitUntilApiTimeoutSeconds: # NOTE: this is the timeout for the single attempt of Cadence/Temporal activity(start to close timeout)
          type: integer
        executeApiTimeoutSeconds: # NOTE: this is the timeout for the single attempt of Cadence/Temporal activity(start to close timeout)
          type: integer
        waitUntilApiRetryPolicy:
          $ref: '#/components/schemas/RetryPolicy'
        executeApiRetryPolicy:
          $ref: '#/components/schemas/RetryPolicy'
        waitUntilApiFailurePolicy:
          $ref: '#/components/schemas/WaitUntilApiFailurePolicy'
        skipWaitUntil:
          type: boolean
        dataAttributesLoadingPolicy:
          $ref: '#/components/schemas/PersistenceLoadingPolicy'
        waitUntilApiDataAttributesLoadingPolicy:
          $ref: '#/components/schemas/PersistenceLoadingPolicy'
        executeApiDataAttributesLoadingPolicy:
          $ref: '#/components/schemas/PersistenceLoadingPolicy'
...
    PersistenceLoadingPolicy:
      type: object
      properties:
        persistenceLoadingType:
          $ref: '#/components/schemas/PersistenceLoadingType'
        partialLoadingKeys:
          type: array
          items:
            type: string
        lockingKeys:
          type: array
          items:
            type: string
        useKeyAsPrefix:
          type: boolean
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
        subStatus:
          $ref: '#/components/schemas/ErrorSubStatus'
        originalWorkerErrorDetail:
          type: string
        originalWorkerErrorType:
          type: string
        originalWorkerErrorStatus:
          type: integer
    ErrorSubStatus:
      type: string
      enum:
        - UNCATEGORIZED_SUB_STATUS
        - WORKFLOW_ALREADY_STARTED_SUB_STATUS
        - WORKFLOW_NOT_EXISTS_SUB_STATUS
        - WORKER_API_ERROR
        - LONG_POLL_TIME_OUT_SUB_STATUS
...
    WorkflowStatus:
      type: string
      enum:
        - RUNNING
        - COMPLETED
        - FAILED
        - TIMEOUT
        - TERMINATED
        - CANCELED
        - CONTINUED_AS_NEW
    StateCompletionOutput:
      type: object
      required:
        - completedStateExecutionId
        - completedStateId
      properties:
        completedStateId:
          type: string
        completedStateExecutionId:
          type: string
        completedStateOutput:
          $ref: '#/components/schemas/EncodedObject'
...
    WorkflowStateStartResponse:
      type: object
      properties:
        localActivityInput: # for debugging succeeded local activity which didn't include input in the history
          type: string
        upsertSearchAttributes:
          type: array
          items:
            $ref: '#/components/schemas/SearchAttribute'
        upsertDataObjects:
          type: array
          items:
            $ref: '#/components/schemas/KeyValue'
        commandRequest:
          $ref: '#/components/schemas/CommandRequest'
        upsertStateLocals:
          type: array
          items:
            $ref: '#/components/schemas/KeyValue'
        recordEvents:
          type: array
          items:
            $ref: '#/components/schemas/KeyValue'
        publishToInterStateChannel:
          type: array
          items:
            $ref: '#/components/schemas/InterStateChannelPublishing'
...
    WorkflowStateDecideResponse:
      type: object
      properties:
        localActivityInput: # for debugging succeeded local activity which didn't include input in the history
          type: string
        stateDecision:
          $ref: '#/components/schemas/StateDecision'
        upsertSearchAttributes:
          type: array
          items:
            $ref: '#/components/schemas/SearchAttribute'
        upsertDataObjects:
          type: array
          items:
            $ref: '#/components/schemas/KeyValue'
        recordEvents:
          type: array
          items:
            $ref: '#/components/schemas/KeyValue'
        upsertStateLocals: # this is probably no much value but we keep it to make the APIs more consistent
          type: array
          items:
            $ref: '#/components/schemas/KeyValue'
        publishToInterStateChannel:
          type: array
          items:
            $ref: '#/components/schemas/InterStateChannelPublishing'
...
    StateMovement:
      type: object
      required:
        - stateId
      properties:
        stateId:
          type: string
        stateInput:
          $ref: '#/components/schemas/EncodedObject'
        stateOptions:
          $ref: '#/components/schemas/WorkflowStateOptions'
        waitForKey:
          type: string
    CommandRequest:
      type: object
      properties:
        deciderTriggerType:
          $ref: '#/components/schemas/DeciderTriggerType'
        commandWaitingType:
          $ref: '#/components/schemas/CommandWaitingType'
        timerCommands:
          type: array
          items:
            $ref: '#/components/schemas/TimerCommand'
        signalCommands:
          type: array
          items:
            $ref: '#/components/schemas/SignalCommand'
        interStateChannelCommands:
          type: array
          items:
            $ref: '#/components/schemas/InterStateChannelCommand'
        commandCombinations:
          type: array
          items:
            $ref: '#/components/schemas/CommandCombination'
...
    CommandResults:
      type: object
      properties:
        signalResults:
          type: array
          items:
            $ref: '#/components/schemas/SignalResult'
        interStateChannelResults:
          type: array
          items:
            $ref: '#/components/schemas/InterStateChannelResult'
        timerResults:
          type: array
          items:
            $ref: '#/components/schemas/TimerResult'
        stateStartApiSucceeded:
          type: boolean
...
    ExecutingStateIdMode:
      type: string
      enum:
        - ENABLED_FOR_ALL
        - ENABLED_FOR_STATES_WITH_WAIT_UNTIL
        - DISABLED
    WorkflowConfig:
      type: object
      properties:
        disableSystemSearchAttribute: # deprecated in favor of executingStateIdMode
          type: boolean
        executingStateIdMode:
          $ref: '#/components/schemas/ExecutingStateIdMode'
        continueAsNewThreshold:
          type: integer
        continueAsNewPageSizeInBytes:
          type: integer
        optimizeActivity:
          type: boolean
        optimizeTimer:
          type: boolean
...
    WorkflowShowRequest:
      type: object
      required:
        - workflowId
      properties:
        workflowId:
          type: string
        runId:
          type: string
    WorkflowShowResponse:
      type: object
      required:
        - input
      properties:
        workflowStartedTimestamp:
          type: integer
          format: int64
        workflowType:
          type: string
        input:
          $ref: '#/components/schemas/InterpreterWorkflowInput'
        continueAsNewSnapshot:
          $ref: '#/components/schemas/ContinueAsNewDumpResponse'
        status:
          $ref: '#/components/schemas/WorkflowStatus'
        historyEvents:
          type: array
          items:
            $ref: '#/components/schemas/IwfHistoryEvent'
    IwfHistoryEventType:
      type: string
      enum: 
        - StateWaitUntil
        - StateExecute
        - RpcExecution
        - SignalReceived
        - WorkflowClosed
    IwfHistoryEvent:
      type: object
      properties:
        eventType:
          $ref: '#/components/schemas/IwfHistoryEventType'
        stateWaitUntil:
          $ref: '#/components/schemas/StateWaitUntilDetails'
        stateExecute:
          $ref: '#/components/schemas/StateExecuteDetails'
        rpcExecution:
          $ref: '#/components/schemas/RpcExecutionDetails'
        signalReceived:
          $ref: '#/components/schemas/SignalReceivedDetails'
        workflowClosed:
          $ref: '#/components/schemas/WorkflowClosedDetails'
    StateWaitUntilDetails:
      type: object
      required:
        - stateExecutionId
        - stateId
        - activityId
      properties:
        stateExecutionId:
          type: string
        stateId:
          type: string
        input:
          $ref: '#/components/schemas/EncodedObject'
        fromEventId:
          type: integer
        stateOptions:
          $ref: '#/components/schemas/WorkflowStateOptions'
        response:
          $ref: '#/components/schemas/WorkflowStateStartResponse'
        firstAttemptStartedTimestamp:
          type: integer
          format: int64
        completedAttemptStartedTimestamp:
          type: integer
          format: int64
        completedAttempt:
          type: integer
        completedTimestamp:
          type: integer
          format: int64
        lastFailedAttemptError:
          type: string
    StateExecuteDetails:
      type: object
      required:
        - stateExecutionId
        - stateId
        - activityId
      properties:
        stateExecutionId:
          type: string
        stateId:
          type: string
        input:
          $ref: '#/components/schemas/EncodedObject'
        fromEventId:
          type: integer
        stateOptions:
          $ref: '#/components/schemas/WorkflowStateOptions'
        stateLocals:
          type: array
          items:
            $ref: '#/components/schemas/KeyValue'
        commandResults:
          $ref: '#/components/schemas/CommandResults'
        response:
          $ref: '#/components/schemas/WorkflowStateDecideResponse'
        firstAttemptStartedTimestamp:
          type: integer
          format: int64
        completedAttemptStartedTimestamp:
          type: integer
          format: int64
        completedAttempt:
          type: integer
        completedTimestamp:
          type: integer
          format: int64
        lastFailedAttemptError:
          type: string
    RpcExecutionDetails:
      type: object
      properties:
        response:
          $ref: '#/components/schemas/ExecuteRpcSignalRequest'
        completedTimestamp:
          type: integer
          format: int64
    SignalReceivedDetails:
      type: object
      properties:
        signalName:
          type: string
        value:
          $ref: '#/components/schemas/EncodedObject'
        completedTimestamp:
          type: integer
          format: int64
    WorkflowClosedDetails:
      type: object
      properties:
        workflowClosedTimestamp:
          type: integer
          format: int64
    # below is additional data models translated from iwf-server
    # https://github.com/indeedeng/iwf/blob/main/service/interfaces.go TODO        
    InterpreterWorkflowInput:
      type: object
      properties:
        iwfWorkflowType:
          type: string
        iwfWorkerUrl:
          type: string
        startStateId:
          type: string
        waitForCompletionStateExecutionIds:
          type: array
          items:
            type: string
        waitForCompletionStateIds:
          type: array
          items:
            type: string
        stateInput:
          $ref: '#/components/schemas/EncodedObject'
        stateOptions:
          $ref: '#/components/schemas/WorkflowStateOptions'
        initSearchAttributes:
          type: array
          items:
            $ref: '#/components/schemas/SearchAttribute'
        initDataAttributes:
          type: array
          items:
            $ref: '#/components/schemas/KeyValue'
        useMemoForDataAttributes:
          type: boolean
        config:
          $ref: '#/components/schemas/WorkflowConfig'
        isResumeFromContinueAsNew:
          type: boolean
          description: "Indicates this is input for continueAsNew. When true, will ignore StartStateId, StateInput, StateOptions, InitSearchAttributes"
        continueAsNewInput:
          $ref: '#/components/schemas/ContinueAsNewInput'
    ContinueAsNewInput:
      type: object
      required:
        - previousInternalRunId
      properties:
        previousInternalRunId:
          type: string
          description: "For loading from previous run"
...
    ExecuteRpcSignalRequest:
      type: object
      properties:
        rpcInput:
          $ref: '#/components/schemas/EncodedObject'
        rpcOutput:
          $ref: '#/components/schemas/EncodedObject'
        upsertDataObjects:
          type: array
          items:
            $ref: '#/components/schemas/KeyValue'
        upsertSearchAttributes:
          type: array
          items:
            $ref: '#/components/schemas/SearchAttribute'
        stateDecision:
          $ref: '#/components/schemas/StateDecision'
        recordEvents:
          type: array
          items:
            $ref: '#/components/schemas/KeyValue'
        interStateChannelPublishing:
          type: array
          items:
            $ref: '#/components/schemas/InterStateChannelPublishing'
...
    InternalTimerStatus:
      type: string
      enum:
        - Pending
        - Fired
        - Skipped
    ContinueAsNewDumpResponse:
      type: object
      properties:
        StatesToStartFromBeginning:
          type: array
          items:
            $ref: '#/components/schemas/StateMovement'
          description: "StatesToStartFromBeginning means they haven't started in the previous run"
        StateExecutionsToResume:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/StateExecutionResumeInfo'
          description: "stateExeId to StateExecutionResumeInfo"
        InterStateChannelReceived:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/EncodedObject'
        SignalsReceived:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/EncodedObject'
        StateExecutionCounterInfo:
          $ref: '#/components/schemas/StateExecutionCounterInfo'
        StateOutputs:
          type: array
          items:
            $ref: '#/components/schemas/StateCompletionOutput'
        StaleSkipTimerSignals:
          type: array
          items:
            $ref: '#/components/schemas/StaleSkipTimerSignal'
        DataObjects:
          type: array
          items:
            $ref: '#/components/schemas/KeyValue'
        SearchAttributes:
          type: array
          items:
            $ref: '#/components/schemas/SearchAttribute'
...
    StateExecutionCounterInfo:
      type: object
      properties:
        stateIdStartedCount:
          type: object
          additionalProperties:
            type: integer
          description: "for stateExecutionId"
        stateIdCurrentlyExecutingCount:
          type: object
          additionalProperties:
            type: integer
          description: "for sys search attribute ExecutingStateIds"
        totalCurrentlyExecutingCount:
          type: integer
          description: "for dead end"
    StateExecutionResumeInfo:
      type: object
      required:
        - stateExecutionId
        - state
        - stateExecutionCompletedCommands
        - commandRequest
        - stateExecutionLocals
      properties:
        stateExecutionId:
          type: string
        state:
          $ref: '#/components/schemas/StateMovement'
        stateExecutionCompletedCommands:
          $ref: '#/components/schemas/StateExecutionCompletedCommands'
        commandRequest:
          $ref: '#/components/schemas/CommandRequest'
        stateExecutionLocals:
          type: array
          items:
            $ref: '#/components/schemas/KeyValue'
    StateExecutionCompletedCommands:
      type: object
      properties:
        completedTimerCommands:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/InternalTimerStatus'
        completedSignalCommands:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/EncodedObject'
        completedInterStateChannelCommands:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/EncodedObject'
    StaleSkipTimerSignal:
      type: object
      properties:
        stateExecutionId:
          type: string
        timerCommandId:
          type: string
        timerCommandIndex:
          type: integer
...